requirements.txt
python-telegram-bot==20.6
python-dotenv==1.0.0
SQLAlchemy==2.0.23
httpx==0.25.0
psycopg2-binary==2.9.9
tabulate==0.9.0
alembic==1.12.1
